move classes de utilidade para .util
